HEEDLS-1006 Fixed being unable to edit delegate details when primary email is a GUID
