+++ DigitalLearningSolutions.Web/Helpers/PromptsService.cs
@@ -8,6 +8,7 @@
     using DigitalLearningSolutions.Data.Services;
     using DigitalLearningSolutions.Web.ViewModels.Common;
     using DigitalLearningSolutions.Web.ViewModels.MyAccount;
+    using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.EditDelegate;
     using Microsoft.AspNetCore.Mvc.ModelBinding;
 
     public class PromptsService
@@ -85,7 +86,7 @@
         }
 
         public List<EditDelegateRegistrationPromptViewModel> GetEditDelegateRegistrationPromptViewModelsForCentre(
-            EditDetailsFormData formData,
+            EditDelegateFormData formData,
             int centreId
         )
         {
@@ -192,6 +193,24 @@
 
         public void ValidateCentreRegistrationPrompts(
             EditDetailsFormData formData,
+            int centreId,
+            ModelStateDictionary modelState
+        )
+        {
+            ValidateCentreRegistrationPrompts(
+                centreId,
+                formData.Answer1,
+                formData.Answer2,
+                formData.Answer3,
+                formData.Answer4,
+                formData.Answer5,
+                formData.Answer6,
+                modelState
+            );
+        }
+
+        public void ValidateCentreRegistrationPrompts(
+            EditDelegateFormData formData,
             int centreId,
             ModelStateDictionary modelState
         )
