+++ DigitalLearningSolutions.Web/ViewModels/Common/EditDetailsFormData.cs
@@ -1,57 +1,15 @@
 ﻿namespace DigitalLearningSolutions.Web.ViewModels.Common
 {
-    using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations;
     using DigitalLearningSolutions.Web.Attributes;
     using DigitalLearningSolutions.Web.Helpers;
 
-    public class EditDetailsFormData
+    public class EditDetailsFormData : EditAccountDetailsFormDataBase
     {
-        [Required(ErrorMessage = "Enter your first name")]
-        [MaxLength(250, ErrorMessage = CommonValidationErrorMessages.TooLongFirstName)]
-        public string? FirstName { get; set; }
-
-        [Required(ErrorMessage = "Enter your last name")]
-        [MaxLength(250, ErrorMessage = CommonValidationErrorMessages.TooLongLastName)]
-        public string? LastName { get; set; }
-
         [Required(ErrorMessage = "Enter your email")]
         [MaxLength(255, ErrorMessage = CommonValidationErrorMessages.TooLongEmail)]
         [EmailAddress(ErrorMessage = CommonValidationErrorMessages.InvalidEmail)]
         [NoWhitespace(CommonValidationErrorMessages.WhitespaceInEmail)]
         public string? Email { get; set; }
-
-        [MaxLength(255, ErrorMessage = CommonValidationErrorMessages.TooLongEmail)]
-        [EmailAddress(ErrorMessage = CommonValidationErrorMessages.InvalidEmail)]
-        [NoWhitespace(CommonValidationErrorMessages.WhitespaceInEmail)]
-        public string? CentreSpecificEmail { get; set; }
-
-        public int? JobGroupId { get; set; }
-
-        public string? Answer1 { get; set; }
-
-        public string? Answer2 { get; set; }
-
-        public string? Answer3 { get; set; }
-
-        public string? Answer4 { get; set; }
-
-        public string? Answer5 { get; set; }
-
-        public string? Answer6 { get; set; }
-
-        public string? ProfessionalRegistrationNumber { get; set; }
-
-        public bool? HasProfessionalRegistrationNumber { get; set; }
-
-        public bool IsSelfRegistrationOrEdit { get; set; }
-
-        public IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
-        {
-            if (!JobGroupId.HasValue)
-            {
-                yield return new ValidationResult("Select a job group", new[] { nameof(JobGroupId) });
-            }
-        }
     }
 }
