+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/EditDelegate/EditDelegateFormData.cs
@@ -7,7 +7,8 @@
     using DigitalLearningSolutions.Web.Helpers;
     using DigitalLearningSolutions.Web.ViewModels.Common;
 
-    public class EditDelegateFormData : EditDetailsFormData, IEditProfessionalRegistrationNumbers, IValidatableObject
+    public class EditDelegateFormData : EditAccountDetailsFormDataBase, IEditProfessionalRegistrationNumbers,
+        IValidatableObject
     {
         public EditDelegateFormData() { }
 
@@ -54,5 +55,7 @@
             HasProfessionalRegistrationNumber = formData.HasProfessionalRegistrationNumber;
             IsSelfRegistrationOrEdit = false;
         }
+
+        public string? Email { get; set; }
     }
 }
